Move service providers to related service definitions
